feat: 3.3.2

+++ src/main/java/com/yeepay/yop/sdk/utils/checksum/CRC64.java
@@ -28,7 +28,7 @@ import java.util.zip.Checksum;
  */
 public class CRC64 implements Checksum {
 
-    private final static long POLY = 0xc96c5795d7870f42L; // ECMA-182
+    private final static long POLY = (long) 0xc96c5795d7870f42L; // ECMA-182
 
     /* CRC64 calculation table. */
     private final static long[] table;
